% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cETL.R
\docType{class}
\name{cETL-class}
\alias{cETL-class}
\alias{cETL}
\title{cETL: Class for DataSet ETL}
\value{
Tables, list, DB
}
\description{
cETL: Class for DataSet ETL
}
\section{Fields}{

\describe{
\item{\code{par}}{data.frame}
}}


