% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cOptions.R
\docType{class}
\name{cOptions-class}
\alias{cOptions-class}
\alias{cOptions}
\title{cOptions: Class for Options Derivatives Analysis}
\value{
Tables, plots
}
\description{
cOptions: Class for Options Derivatives Analysis
}
\section{Fields}{

\describe{
\item{\code{par}}{data.frame.}
}}


