% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cPortfolio.R
\docType{class}
\name{cPortfolio-class}
\alias{cPortfolio-class}
\alias{cPortfolio}
\title{cPortfolio: Class for Portfolio Management}
\value{
Tables, plots
}
\description{
cPortfolio: Class for Portfolio Management
}
\section{Fields}{

\describe{
\item{\code{par}}{data.frame.}
}}


