% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTimeSeries.R
\docType{class}
\name{cTimeSeries-class}
\alias{cTimeSeries-class}
\alias{cTimeSeries}
\title{cTimeSeries: Class for TimeSeries Analysis}
\value{
Tables, plots
}
\description{
cTimeSeries: Class for TimeSeries Analysis
}
\section{Fields}{

\describe{
\item{\code{par}}{data.frame.}
}}


