% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DataSet.R
\name{mod_DataSet_server}
\alias{mod_DataSet_server}
\title{All server side functions of DataSet Button}
\usage{
mod_DataSet_server(id, up_input, .e = .e)
}
\arguments{
\item{id}{NS(id)}

\item{up_input}{server side inputs}

\item{.e}{.e env}
}
\value{
DataSet server
}
\description{
All server side functions of DataSet Button
}
