% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DataSet.R
\name{mod_DataSet_ui}
\alias{mod_DataSet_ui}
\title{All ui functions of DataSet Button}
\usage{
mod_DataSet_ui(id, configs, .e = .e)
}
\arguments{
\item{id}{NS(id)}

\item{configs}{configs}

\item{.e}{.e env}
}
\value{
DataSet ui
}
\description{
All ui functions of DataSet Button
}
