% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Derivatives_Options.R
\name{mod_Derivatives_Options_server}
\alias{mod_Derivatives_Options_server}
\title{All server side functions of Derivatives / Options Tab}
\usage{
mod_Derivatives_Options_server(input, output, session, id, up_input, .e = .e)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{id}{NS(id)}

\item{up_input}{server side inputs}

\item{.e}{.e env}
}
\value{
Derivatives / Options Analysis
}
\description{
All server side functions of Derivatives / Options Tab
}
