% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Derivatives_Options.R
\name{mod_Derivatives_Options_ui}
\alias{mod_Derivatives_Options_ui}
\title{All ui and server of Derivatives / Options Tab}
\usage{
mod_Derivatives_Options_ui(id, configs)
}
\arguments{
\item{id}{NS(id)}

\item{configs}{configs}
}
\value{
Derivatives / Options ui
}
\description{
All ui and server of Derivatives / Options Tab
}
