% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Derivatives_Screener.R
\name{mod_Derivatives_Screener_ui}
\alias{mod_Derivatives_Screener_ui}
\title{All ui and server of Derivatives / Screener Tab}
\usage{
mod_Derivatives_Screener_ui(id, configs)
}
\arguments{
\item{id}{NS(id)}

\item{configs}{configs}
}
\value{
Derivatives / Screener ui
}
\description{
All ui and server of Derivatives / Screener Tab
}
