% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Derivatives.R
\name{mod_Derivatives_server}
\alias{mod_Derivatives_server}
\title{All server side functions of Derivatives Tab}
\usage{
mod_Derivatives_server(id, up_input, .e = .e)
}
\arguments{
\item{id}{NS(id)}

\item{up_input}{server side inputs}

\item{.e}{.e env}
}
\value{
Derivatives server
}
\description{
All server side functions of Derivatives Tab
}
