% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Derivatives.R
\name{mod_Derivatives_ui}
\alias{mod_Derivatives_ui}
\title{All ui and server of Derivatives tab}
\usage{
mod_Derivatives_ui(id, configs)
}
\arguments{
\item{id}{NS(id)}

\item{configs}{configs}
}
\value{
Derivatives Analysis
}
\description{
All ui and server of Derivatives tab
}
