% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MainBox.R
\name{mod_MainBox_ui}
\alias{mod_MainBox_ui}
\title{Title mod_MainBox ui & server}
\usage{
mod_MainBox_ui(id, params = NULL, mainui = NULL)
}
\arguments{
\item{id}{NS(id)}

\item{params}{Parameters}

\item{mainui}{contents of Box / bs4Card}
}
\value{
bs4dash Box
}
\description{
mod_MainBox_ui(id, ...)
mod_MainBox_server(id, ...)
}
