% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MarketScanner.R
\name{mod_MarketScanner_ui}
\alias{mod_MarketScanner_ui}
\title{Title All ui and server of Market Scanner tab}
\usage{
mod_MarketScanner_ui(id, .e = .e)
}
\arguments{
\item{id}{NS(id)}

\item{.e}{.e env}
}
\value{
Market info
}
\description{
Title All ui and server of Market Scanner tab
}
