% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Portfolio_Dependency.R
\name{mod_Portfolio_Dependency_server}
\alias{mod_Portfolio_Dependency_server}
\title{All server side functions of Portfolio / Dependency Structure  Tab}
\usage{
mod_Portfolio_Dependency_server(input, output, session, id, up_input, .e = .e)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{id}{NS(id)}

\item{up_input}{server side inputs}

\item{.e}{.e env}
}
\value{
Portfolio / Dependency Structure  Analysis
}
\description{
All server side functions of Portfolio / Dependency Structure  Tab
}
