% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Portfolio_Dependency.R
\name{mod_Portfolio_Dependency_ui}
\alias{mod_Portfolio_Dependency_ui}
\title{All ui and server of Portfolio / Dependency Structure Tab}
\usage{
mod_Portfolio_Dependency_ui(id, configs)
}
\arguments{
\item{id}{NS(id)}

\item{configs}{configs}
}
\value{
Portfolio / Dependency Structure ui
}
\description{
All ui and server of Portfolio / Dependency Structure Tab
}
