% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Portfolio_Optimization.R
\name{mod_Portfolio_Optimization_server}
\alias{mod_Portfolio_Optimization_server}
\title{All server side functions of Portfolio / Optimization Tab}
\usage{
mod_Portfolio_Optimization_server(
  input,
  output,
  session,
  id,
  up_input,
  .e = .e
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{id}{NS(id)}

\item{up_input}{server side inputs}

\item{.e}{.e env}
}
\value{
Portfolio / Optimization Analysis
}
\description{
All server side functions of Portfolio / Optimization Tab
}
