% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Portfolio_Optimization.R
\name{mod_Portfolio_Optimization_ui}
\alias{mod_Portfolio_Optimization_ui}
\title{All ui and server of Portfolio / Optimization Tab}
\usage{
mod_Portfolio_Optimization_ui(id, configs, .e = .e)
}
\arguments{
\item{id}{NS(id)}

\item{configs}{configs}

\item{.e}{.e env}
}
\value{
Portfolio / Optimization ui
}
\description{
All ui and server of Portfolio / Optimization Tab
}
