% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Portfolio_RiskReturn.R
\name{mod_Portfolio_RiskReturn_ui}
\alias{mod_Portfolio_RiskReturn_ui}
\title{All ui and server of Portfolio / RiskReturn Tab}
\usage{
mod_Portfolio_RiskReturn_ui(id, configs, .e = .e)
}
\arguments{
\item{id}{NS(id)}

\item{configs}{configs}

\item{.e}{.e env}
}
\value{
Portfolio / RiskReturn ui
}
\description{
All ui and server of Portfolio / RiskReturn Tab
}
