% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Portfolio.R
\name{mod_Portfolio_server}
\alias{mod_Portfolio_server}
\title{All server side functions of Portfolio Tab}
\usage{
mod_Portfolio_server(id, up_input, .e = .e)
}
\arguments{
\item{id}{NS(id)}

\item{up_input}{server side inputs}

\item{.e}{.e env}
}
\value{
Portfolio server
}
\description{
All server side functions of Portfolio Tab
}
