% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Portfolio.R
\name{mod_Portfolio_ui}
\alias{mod_Portfolio_ui}
\title{All ui and server of Portfolio tab}
\usage{
mod_Portfolio_ui(id, configs, .e = .e)
}
\arguments{
\item{id}{NS(id)}

\item{configs}{configs}

\item{.e}{.e env}
}
\value{
Portfolio Analysis
}
\description{
All ui and server of Portfolio tab
}
